Split message structs into JSON and RPC structs.
Before I was using the same set of structs for both JSON and go RPC.
Since we will be encoding times and durations differently for RPC and
because of the way RPC handles pointers to zero values of primitive
types differently than JSON, its time to have two sets of structs.

I will need to add tests for the new RPC structs later, but I wanted to
get something out there for you to look at.