Subtract released memory from /proc/memory/total built-in metric.
runtime.MemStats.Sys contains the total virtual address space used by the
process. It does not represent the amount of physical memory used by the
process. System memory that was allocated and later released is included in
this value, which is not intuitive. This should probably have been called
/proc/memory/virtual.

By subtracting runtime.MemStats.HeapReleased (which contains the amount of
physical memory released to the OS), we get a much better approximation of
system (physical) memory currently used.